# use type() in int validator 