@samuelcolvin 2018-09-20
Looks good, my only question is: what happens with the following:

```py
class MyInt(int):
    pass

class FooModel(BaseModel):
   v: MyInt
```

I would expect it to be validated as an int. More dangerous people could already have code like this.

Maybe safter to have:

```py
def int_validator(v) -> int:
    if not isinstance(v, bool) and isinstance(v, int):
        return v
```
?